chore(api): move repository endpoints into separate module
